Eliminated floating point values from replays for portability
Player power is now internally an integer. There is no good reason to use floating point math there. Stored as an uint16 in replays.
Player coordinates are now floored when a stage starts. This is unnoticable to the player. Stored as a pair of uint16s in replays.

+++ src/player.c
@@ -38,7 +38,7 @@ Animation *player_get_ani(Character cha) {
 	return ani;
 }
 
-void player_set_power(Player *plr, float npow) {
+void player_set_power(Player *plr, short npow) {
 	switch(plr->cha) {
 		case Youmu:
 			youmu_power(plr, npow);
