+++ src/plrmodes.h
@@ -21,13 +21,13 @@ int youmu_homing(Projectile *p, int t);
 
 void youmu_shot(Player *plr);
 void youmu_bomb(Player *plr);
-void youmu_power(Player *plr, float npow);
+void youmu_power(Player *plr, short npow);
 
 /* Marisa */
 
 void marisa_shot(Player *plr);
 void marisa_bomb(Player *plr);
-void marisa_power(Player *plr, float npow);
+void marisa_power(Player *plr, short npow);
 
 /* Misc */
 int plrmode_repr(char *out, size_t outsize, Character pchar, ShotMode pshot);
